Updated XSD and Link object to include the detail_level.  Updated the NetworkTest to handle this as a Long rather than an int. Changes to be committed: modified:   src/main/java/edu/berkeley/path/model_objects/network/Link.java modified:   src/main/resources/network.xsd modified:   src/test/java/edu/berkeley/path/model_objects/network/NetworkTest.java

+++ src/test/java/edu/berkeley/path/model_objects/network/NetworkTest.java
@@ -68,7 +68,7 @@ public class NetworkTest {
     ln.setLanes(4.0d);
     ln.setLength(1000.0d);
     ln.setWKT("TEST WKT");
-    ln.setDetailLevel(5);
+    ln.setDetailLevel(5L);
     
     ln.setBeginNode(nd1);
     ln.setEndNode(nd2);
@@ -107,7 +107,7 @@ public class NetworkTest {
     assertEquals("three", ln3.getFirstRoadName() );
 		assertEquals(2, ln3.getPoints().size());
     assertEquals("TEST WKT", ln3.getWKT());
-    assertEquals(5, ln3.getDetailLevel());
+    assertEquals(5L, ln3.getDetailLevel().longValue());
     
     // note that lookups by string ID work too:
     Node n2 = nw.getNodeWithId(2);
